add /fsr, /eval, and FACT_SOURCE override envar (#88)
* add /fsr, /eval, and FACT_SOURCE

* remove duplicate codepaths, add missing imports

* remove unused imports

* use full_text instead of the blob
